fix: folder renaming error correction
