+++ src/widgets/CartPreview/ui/CartPreview.module.css
@@ -6,19 +6,9 @@
   border: 1px solid var(--main-color-4);
 }
 
-.productList {
-  list-style: none;
-  counter-reset: li;
-}
-
 .productCart {
   display: flex;
   width: 100%;
   gap: 10px;
   align-items: start;
-}
-
-.productCart::before {
-  counter-increment: li;
-  content: counters(li, '.') '. ';
 }
